little refactoring in battle menus
