+++ Core/Battle/Menu/IGMDataItem/Gradient/Renzokeken.cs
@@ -8,20 +8,31 @@ namespace OpenVIII.IGMDataItem.Gradient
     {
         #region Fields
 
+        protected bool _trigger;
         protected Color Color_default;
         protected Slide<int> HitSlide;
         protected Rectangle HotSpot;
-        protected bool _trigger;
 
         #endregion Fields
 
         #region Constructors
-        public override void Reset()
+
+        private Renzokeken()
         {
-            Restart();
-            base.Reset();
         }
-        private Renzokeken() { }
+
+        #endregion Constructors
+
+        #region Properties
+
+        public bool Done => HitSlide.Done || !Enabled;
+
+        public bool Trigger { get => (_trigger && !Done); private set => _trigger = value; }
+
+        #endregion Properties
+
+        #region Methods
+
         public static Renzokeken Create(Rectangle? pos = null, Color? color = null, Color? faded_color = null, float blink_adjustment = 1f, Rectangle? hotspot = null, Rectangle? restriction = null, double time = 0d, double delay = 0d, Color? darkcolor = null, bool rev = false, bool vanish = true)
         {
             int total = 12 + 180;
@@ -72,17 +83,11 @@ namespace OpenVIII.IGMDataItem.Gradient
             return r;
         }
 
-        #endregion Constructors
-
-        #region Properties
-
-        public bool Trigger { get => (_trigger && !Done); private set => _trigger = value; }
-        public bool Done => HitSlide.Done || !Enabled;
-
-        
-        #endregion Properties
-
-        #region Methods
+        public override void Reset()
+        {
+            Restart();
+            base.Reset();
+        }
 
         public void Restart()
         {
